--- api/src/routes/cards.ts
@@ -1,35 +1,0 @@
-/* eslint-disable @typescript-eslint/no-misused-promises */
-import { Request, Response, NextFunction, Router } from 'express'
-import db from './../db'
-const { Card } = db
-
-const cardsRoute = Router()
-
-interface CardAttributes {
-  id: number
-  name: string
-  dmg: number
-  life: string
-  ability: string
-  abilities: string[]
-  race: string
-  cost: number
-  canAttack: 'enum'
-  flying: boolean
-}
-
-const getAllCards = async (): Promise<CardAttributes[]> => {
-  const allPacksDB = await Card.findAll()
-  return allPacksDB
-}
-
-cardsRoute.get('/', async (_req: Request, res: Response, next: NextFunction) => {
-  try {
-    const packs = await getAllCards()
-    return res.send(packs)
-  } catch (error) {
-    return next(error)
-  }
-})
-
-export default cardsRoute
