+++ api/src/seeders/starsPack.js
